+++ tango/workspaces/remote_workspace.py
@@ -1,25 +1,26 @@
-import json
 import logging
-import random
 import tempfile
 import warnings
 from abc import abstractmethod
-from collections import OrderedDict
 from contextlib import contextmanager
 from pathlib import Path
-from typing import Dict, Generator, Iterable, Iterator, Optional, TypeVar, Union, cast
+from typing import (
+    Dict,
+    Generator,
+    Iterable,
+    Iterator,
+    Optional,
+    Tuple,
+    TypeVar,
+    Union,
+)
 from urllib.parse import ParseResult
-
-import petname
 
 from tango.common.exceptions import StepStateError
 from tango.common.logging import file_handler
 from tango.common.remote_utils import (
     RemoteClient,
     RemoteConstants,
-    RemoteDataset,
-    RemoteDatasetConflict,
-    RemoteDatasetNotFound,
     RemoteStepLock,
 )
 from tango.common.util import exception_to_string, tango_cache_dir, utc_now_datetime
@@ -44,9 +45,6 @@ class RemoteWorkspace(Workspace):
     Constants = RemoteConstants
     STEP_INFO_CACHE_SIZE: int = 512
     NUM_CONCURRENT_WORKERS: int = 9
-
-    def __init__(self):
-        super().__init__()
 
     @property
     @abstractmethod
@@ -195,33 +193,12 @@ class RemoteWorkspace(Workspace):
         finally:
             self.locks.pop(step).release()
 
-    def register_run(self, targets: Iterable[Step], name: Optional[str] = None) -> Run:
+    def _get_run_step_info(self, targets: Iterable[Step]) -> Tuple[Dict, Dict]:
         import concurrent.futures
 
         all_steps = set(targets)
         for step in targets:
             all_steps |= step.recursive_dependencies
-
-        # Create a remote dataset that represents this run. The dataset which just contain
-        # a JSON file that maps step names to step unique IDs.
-        run_dataset: RemoteDataset
-        if name is None:
-            # Find a unique name to use.
-            while True:
-                name = petname.generate() + str(random.randint(0, 100))
-                try:
-                    run_dataset = self.client.create(
-                        self.Constants.run_dataset_name(cast(str, name))
-                    )
-                except RemoteDatasetConflict:
-                    continue
-                else:
-                    break
-        else:
-            try:
-                run_dataset = self.client.create(self.Constants.run_dataset_name(name))
-            except RemoteDatasetConflict:
-                raise ValueError(f"Run name '{name}' is already in use")
 
         steps: Dict[str, StepInfo] = {}
         run_data: Dict[str, str] = {}
@@ -241,51 +218,28 @@ class RemoteWorkspace(Workspace):
                 steps[step_info.step_name] = step_info
                 run_data[step_info.step_name] = step_info.unique_id
 
-        # Upload run data to dataset.
-        # NOTE: We don't commit the dataset here since we'll need to upload the logs file
-        # after the run.
-        self.client.upload(
-            run_dataset, json.dumps(run_data).encode(), self.Constants.RUN_DATA_FNAME
-        )
+        return steps, run_data
 
-        return Run(name=cast(str, name), steps=steps, start_date=run_dataset.created)
+    @abstractmethod
+    def _save_run(
+        self, steps: Dict[str, StepInfo], run_data: Dict[str, str], name: Optional[str] = None
+    ) -> Run:
+        raise NotImplementedError()
+
+    def register_run(self, targets: Iterable[Step], name: Optional[str] = None) -> Run:
+        steps, run_data = self._get_run_step_info(targets)
+        run = self._save_run(steps, run_data, name)
+        return run
 
     def registered_runs(self) -> Dict[str, Run]:
-        import concurrent.futures
-
-        runs: Dict[str, Run] = {}
-
-        with concurrent.futures.ThreadPoolExecutor(
-            max_workers=self.NUM_CONCURRENT_WORKERS,
-            thread_name_prefix="RemoteWorkspace.registered_runs()-",
-        ) as executor:
-            run_futures = []
-            for dataset in self.client.datasets(match=self.Constants.RUN_DATASET_PREFIX):
-                run_futures.append(executor.submit(self._get_run_from_dataset, dataset))
-            for future in concurrent.futures.as_completed(run_futures):
-                run = future.result()
-                if run is not None:
-                    runs[run.name] = run
-
-        return runs
+        raise NotImplementedError()
 
     def registered_run(self, name: str) -> Run:
-        err_msg = f"Run '{name}' not found in workspace"
+        raise NotImplementedError()
 
-        try:
-            dataset_for_run = self.client.get(self.Constants.run_dataset_name(name))
-            # TODO: what's this check? seems to be beaker specific.
-            # # Make sure the run is in our workspace.
-            # if dataset_for_run.workspace_ref.id != self.beaker.workspace.get().id:
-            #     raise DatasetNotFound
-        except RemoteDatasetNotFound:
-            raise KeyError(err_msg)
-
-        run = self._get_run_from_dataset(dataset_for_run)
-        if run is None:
-            raise KeyError(err_msg)
-        else:
-            return run
+    @abstractmethod
+    def _save_run_log(self, name: str, log_file: Path):
+        raise NotImplementedError()
 
     @contextmanager
     def capture_logs_for_run(self, name: str) -> Generator[None, None, None]:
@@ -295,52 +249,8 @@ class RemoteWorkspace(Workspace):
                 with file_handler(log_file):
                     yield None
             finally:
-                run_dataset = self.Constants.run_dataset_name(name)
-                self.client.sync(run_dataset, log_file)
-                self.client.commit(run_dataset)
+                self._save_run_log(name, log_file)
 
-    def _get_run_from_dataset(self, dataset: RemoteDataset) -> Optional[Run]:
-        if dataset.name is None:
-            return None
-        if not dataset.name.startswith(self.Constants.RUN_DATASET_PREFIX):
-            return None
-
-        try:
-            run_name = dataset.name[len(self.Constants.RUN_DATASET_PREFIX) :]
-            steps: Dict[str, StepInfo] = {}
-            steps_info_bytes = self.client.get_file(dataset, self.Constants.RUN_DATA_FNAME)
-            steps_info = json.loads(steps_info_bytes)
-        except (RemoteDatasetNotFound, FileNotFoundError):
-            return None
-
-        import concurrent.futures
-
-        with concurrent.futures.ThreadPoolExecutor(
-            max_workers=self.NUM_CONCURRENT_WORKERS,
-            thread_name_prefix="RemoteWorkspace._get_run_from_dataset()-",
-        ) as executor:
-            step_info_futures = []
-            for unique_id in steps_info.values():
-                step_info_futures.append(executor.submit(self.step_info, unique_id))
-            for future in concurrent.futures.as_completed(step_info_futures):
-                step_info = future.result()
-                assert step_info.step_name is not None
-                steps[step_info.step_name] = step_info
-
-        return Run(name=run_name, start_date=dataset.created, steps=steps)
-
+    @abstractmethod
     def _update_step_info(self, step_info: StepInfo):
-        dataset_name = self.Constants.step_dataset_name(step_info)
-
-        step_info_dataset: RemoteDataset
-        try:
-            self.client.create(dataset_name)
-        except RemoteDatasetConflict:
-            pass
-        step_info_dataset = self.client.get(dataset_name)
-
-        self.client.upload(
-            step_info_dataset,  # folder name
-            json.dumps(step_info.to_json_dict()).encode(),  # step info dict.
-            self.Constants.STEP_INFO_FNAME,  # step info filename
-        )
+        raise NotImplementedError()
